refactor: RestaurantList 관련 모듈 폴더 구조 수정
